can edit helmets with list
